Move quic_logging_impl.{h,cc} to quiche_logging_impl.{h,cc} to match the name expected by the QUICHE platform API
