+++ source/common/quic/platform/quiche_bug_tracker_impl.cc
@@ -1,7 +1,7 @@
 #include "source/common/quic/platform/quiche_bug_tracker_impl.h"
 
 #include "source/common/common/assert.h"
-#include "source/common/quic/platform/quic_logging_impl.h"
+#include "source/common/quic/platform/quiche_logging_impl.h"
 
 // NOLINT(namespace-envoy)
 //
